Add new interconnect proto
The Interconnect API is intended to enable a network orchestrator to use connectivity services from multiple connectivity service providers without integrating with each service provider separately. In particular the API is intended to facilitate use of connectivity services that require the customer to establish a single wireless link with the service provider’s network, through which connectivity is available. This wireless link could be a link to an optical terminal on a low-Earth-orbiting satellite, or it could be a link to a ground station.